Substituted class UserDAOSQLite with class UserDAOSQLiteWithUserMapper which uses UserMapper. Added UserMapper.
